iter

+++ lucene/core/src/java/org/apache/lucene/util/packed/DirectMonotonicReader.java
@@ -143,7 +143,7 @@ public final class DirectMonotonicReader extends LongValues implements Accountab
 
   /** Get lower/upper bounds for the value at a given index without hitting the direct reader. */
   private long[] getBounds(long index) {
-    final int block = (int) (index >>> blockShift);
+    final int block = Math.toIntExact(index >>> blockShift);
     final long blockIndex = index & ((1 << blockShift) - 1);
     final long lowerBound = mins[block] + (long) (avgs[block] * blockIndex);
     final long upperBound = lowerBound + (1L << bpvs[block]) - 1;
